feat(decorator pattern): change the interface to sync with typescript contract
This commits add a new way to use the decorators. In this way is possible to use method decorators
that already exists in functions. If you write a decorator for a function you can use in methods
too.

BREAKING CHANGE: All contract type was changed. The old version wont work anymore